0.2.2 - no alert on service commands - exception handling - using fieldset - max seconds to 600 - some more optimization

+++ routes/service.js
@@ -1,41 +1,79 @@
 const express = require('express');
 const router = express.Router();
 const shell = require('shelljs');
+const isWin = process.platform === "win32";
 
 router.get('/restart', function (req, res, next) {
+  if (isWin) {
+    res.status(204).send();
+    return;
+  }
+
   shell.exec('systemctl restart acc-server.service', {
     silent: true
   });
-  res.set('Cache-Control', 'no-store').status(200).send();
+
+  res.set('Cache-Control', 'no-cache, no-store, must-revalidate')
+    .status(200).send();
 });
 
 router.get('/start', function (req, res, next) {
+  if (isWin) {
+    res.status(204).send();
+    return;
+  }
+
   shell.exec('systemctl start acc-server.service', {
     silent: true
   });
-  res.set('Cache-Control', 'no-store').status(200).send();
+
+  res.set('Cache-Control', 'no-cache, no-store, must-revalidate')
+    .status(200).send();
 });
 
 router.get('/stop', function (req, res, next) {
+  if (isWin) {
+    res.status(204).send();
+    return;
+  }
+
   shell.exec('systemctl stop acc-server.service', {
     silent: true
   });
-  res.set('Cache-Control', 'no-store').status(200).send();
+
+  res.set('Cache-Control', 'no-cache, no-store, must-revalidate')
+    .status(200).send();
 });
 
 router.get('/status', function (req, res, next) {
+  if (isWin) {
+    res.json({
+      status: ''
+    });
+    return;
+  }
+
   const status = shell.exec('systemctl status acc-server.service | grep -Po \'(?<=Active: ).*\'', {
     silent: true
   }).stdout.replace(/(\r\n|\n|\r)/gm, '');
+
   res.json({
     status
   });
 });
 
 router.get('/log', function (req, res, next) {
+  if (isWin) {
+    res.json({
+      logs: ''
+    });
+    return;
+  }
+
   const logs = shell.exec('journalctl -u acc-server.service -n 1999 | grep -vP \'not enough CPU power|-- Logs begin at\'', {
     silent: true
   }).stdout.replace(/(\r\n|\n|\r)/gm, '</br>');
+
   res.json({
     logs
   });
